fix(396/§2.6): a2r_std time now_ms/now_sec i32→i64——与 stdlib 声明对齐
auto-ai-client daemon.at 重生成时 4 处 E0308：a2r 推断
let deadline: i64 = time.now_ms() + 3000，而 a2r_std::time::now_ms 返回
i32。根因是 a2r-std/src/time.rs 手抄滞后——stdlib/auto/time.rs.at 与
time.vm.at 均声明 i64。i32 截断本身即潜伏 bug（epoch 毫秒 ~1.8e12 每
~24.8 天回绕一次，超时比较在回绕点失准）。

无仓内依赖者（a2r-std 独立 crate）；golden 不涉及 now_ms/now_sec；
auto-cosmic 的 now_secs(u64) 是无关 ClockPort 方法。

+++ crates/a2r-std/src/time.rs
@@ -3,20 +3,24 @@
 
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
-/// Get current time in milliseconds since Unix epoch (truncated to i32)
-pub fn now_ms() -> i32 {
+/// Get current time in milliseconds since Unix epoch
+/// Plan 396 §2.6: i64 per stdlib/auto/time.rs.at (`now_ms() i64`) and
+/// time.vm.at — the old i32 truncation wrapped epoch ms every ~24.8 days
+/// and E0308'd against a2r's i64 local inference (auto-ai-client daemon).
+pub fn now_ms() -> i64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_millis() as i32
+        .as_millis() as i64
 }
 
-/// Get current time in seconds since Unix epoch (truncated to i32)
-pub fn now_sec() -> i32 {
+/// Get current time in seconds since Unix epoch
+/// Plan 396 §2.6: i64 per stdlib/auto/time.rs.at (`now_sec() i64`).
+pub fn now_sec() -> i64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_secs() as i32
+        .as_secs() as i64
 }
 
 /// Sleep for specified milliseconds
